+++ x-pack/plugins/alerting/common/bulk_edit.ts
@@ -1,0 +1,16 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import { Rule } from './rule';
+
+export type BulkEditSkipReason = 'RULE_NOT_MODIFIED';
+
+export interface BulkActionSkipResult {
+  id: Rule['id'];
+  name?: Rule['name'];
+  skip_reason: BulkEditSkipReason;
+}
