+++ x-pack/plugins/security_solution/common/detection_engine/rule_management/api/rules/bulk_actions/response_schema.ts
@@ -5,7 +5,9 @@
  * 2.0.
  */
 
+import type { Rule, BulkActionSkipResult } from '@kbn/alerting-plugin/common';
 import type { BulkActionsDryRunErrCode } from '../../../../../constants';
+
 export interface RuleDetailsInError {
   id: string;
   name?: string;
@@ -15,4 +17,26 @@ export interface NormalizedRuleError {
   status_code: number;
   err_code?: BulkActionsDryRunErrCode;
   rules: RuleDetailsInError[];
+}
+export interface BulkEditActionResults {
+  updated: Rule[];
+  created: Rule[];
+  deleted: Rule[];
+  skipped: BulkActionSkipResult[];
+}
+
+export interface BulkEditActionSummary {
+  failed: number;
+  skipped: number;
+  succeeded: number;
+  total: number;
+}
+export interface BulkEditActionResponse {
+  success?: boolean;
+  rules_count?: number;
+  attributes: {
+    results: BulkEditActionResults;
+    summary: BulkEditActionSummary;
+    errors?: NormalizedRuleError[];
+  };
 }
