+++ x-pack/plugins/security_solution/server/lib/detection_engine/rule_management/api/rules/bulk_actions/route.ts
@@ -16,14 +16,9 @@ import type {
 } from '@kbn/core/server';
 
 import type { RulesClient, BulkOperationError } from '@kbn/alerting-plugin/server';
-import type { SanitizedRule } from '@kbn/alerting-plugin/common';
+import type { SanitizedRule, BulkActionSkipResult } from '@kbn/alerting-plugin/common';
 import { AbortError } from '@kbn/kibana-utils-plugin/common';
-import type {
-  BulkActionSkipResult,
-  BulkEditActionResults,
-} from '@kbn/alerting-plugin/server/rules_client';
 import type { RuleAlertType, RuleParams } from '../../../../rule_schema';
-
 import type { BulkActionsDryRunErrCode } from '../../../../../../../common/constants';
 import {
   DETECTION_ENGINE_RULES_BULK_ACTION,
@@ -38,6 +33,7 @@ import {
 import type {
   NormalizedRuleError,
   RuleDetailsInError,
+  BulkEditActionResponse,
 } from '../../../../../../../common/detection_engine/rule_management/api/rules/bulk_actions/response_schema';
 import type { SetupPlugins } from '../../../../../../plugin';
 import type { SecuritySolutionPluginRouter } from '../../../../../../types';
@@ -73,22 +69,6 @@ export type BulkActionError =
   | PromisePoolError<string>
   | PromisePoolError<RuleAlertType>
   | BulkOperationError;
-
-export interface BulkEditActionSummary {
-  failed: number;
-  skipped: number;
-  succeeded: number;
-  total: number;
-}
-export interface BulkEditActionResponse {
-  success?: boolean;
-  rules_count?: number;
-  attributes: {
-    results: BulkEditActionResults;
-    summary: BulkEditActionSummary;
-    errors?: NormalizedRuleError[];
-  };
-}
 
 const normalizeErrorResponse = (errors: BulkActionError[]): NormalizedRuleError[] => {
   const errorsMap = new Map<string, NormalizedRuleError>();
